Make mid-game dropping more reliable (and significant optimizations) (#136)
* some stuff!

* Comment out a test.

* Optimize game data cache. indexOf takes about 80% less time on average.

* Remove log lines and locks

* Improve accuracy of ping measurement and calculations

+++ emulinker/src/main/java/org/emulinker/kaillera/controller/v086/action/ACKAction.kt
@@ -83,13 +83,7 @@ class ACKAction : V086Action<ClientAck>, V086UserEventHandler<UserEvent> {
       for (user in server.usersMap.values) {
         if (user.status != UserStatus.CONNECTING && user != thisUser)
           users.add(
-            ServerStatus.User(
-              user.name!!,
-              user.ping.toLong(),
-              user.status,
-              user.id,
-              user.connectionType
-            )
+            ServerStatus.User(user.name!!, user.ping, user.status, user.id, user.connectionType)
           )
       }
     } catch (e: MessageFormatException) {
